Update Soft-Real-Time-Device.md
Add note to clarify that the real-time core indices will remain fixed if increasing the CPU core count afterwards. This means that there's no guarantee that the real-time cores will reside on the the last core indices. This is at least observed behavior on Windows 10 IoT enterprise LTSC 2021.

Example to illustrate:
Create a VM with 4 CPU cores
Configure 2 real-time cores.
Observe that the 3rd & 4th core (out of 4) is assigned real-time.
Increase the CPU core count to 6.
Observe that the 3rd & 4th core (out of 6) is still assigned real-time, whereas the last two cores behave regularly.